Several improvements on Alpha strong and quark masses
NOTICE: Now by default the input top mass is the MSbar one. For compatibility it is still possible to use the pole mass by setting FlagMtPole to true (deprecated).
Added method to compute m(m) from mpole using the inversion instead of the iterative method. By default this is not used, but it can be used by setting FlagMpole2MbarNumeric to false.